daemon/requirements.txt
cryptography>=37.0.4
PyKCS11>=1.5.10
pycrypto>=2.6.1
requests>=2.28.1
scandir>=1.10.0
selenium>=4.3.0
websockets>=10.3
webdriver-manager>=3.8.2 